--- service/src/main/java/net/softel/ai/classify/ObjectDetectionController.java
@@ -1,30 +1,0 @@
-// /*
-//  * Copyright 2020 Amazon.com, Inc. or its affiliates. All Rights Reserved.
-//  *
-//  * Licensed under the Apache License, Version 2.0 (the "License"). You may not use this file except in compliance
-//  * with the License. A copy of the License is located at
-//  *
-//  * http://aws.amazon.com/apache2.0/
-//  *
-//  * or in the "license" file accompanying this file. This file is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES
-//  * OR CONDITIONS OF ANY KIND, either express or implied. See the License for the specific language governing permissions
-//  * and limitations under the License.
-//  */
-// package net.softel.ai.classify;
-
-// import org.springframework.stereotype.Controller;
-// import org.springframework.ui.Model;
-// import org.springframework.web.bind.annotation.GetMapping;
-// import org.springframework.web.bind.annotation.RequestParam;
-
-// //@TODO delete this
-// @Controller
-// public class ObjectDetectionController {
-
-//     @GetMapping("/object-detection")
-//     public String greeting(@RequestParam(name="name", required=false, defaultValue="World") String name, Model model) {
-//         model.addAttribute("name", name);
-//         return "object-detection";
-//     }
-
-// }
