+++ service/src/main/java/net/softel/ai/classify/controller/TrainingController.java
@@ -9,8 +9,6 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 import org.springframework.beans.factory.annotation.Qualifier;
 
-// import io.swagger.v3.oas.annotations.Operation;
-// import io.swagger.v3.oas.annotations.tags.Tag;
 import lombok.extern.slf4j.Slf4j;
 import net.softel.ai.classify.train.Training;
 
@@ -22,7 +20,6 @@ import javax.validation.Valid;
 @Slf4j
 @RestController
 @RequestMapping("/train")
-// @Tag(name = "Train", description = "Slack request and callbacks")
 public class TrainingController {
 
     @Autowired
@@ -32,21 +29,10 @@ public class TrainingController {
     @Qualifier("imageClassifierTrainer")
     Training imageClassifierTrainer;
 
-
-    //@Autowired
-    //@Qualifier("imageClassifierTrainer")
-    //Training detectorTrainer;
-
-    @PostMapping(path="/classifier", produces = "application/json")
+    @PostMapping(path="/image/classifier", produces = "application/json")
     public ResponseEntity<String> trainClassifierAsync(@RequestBody @Valid TrainingSuite suite){ 
          trainService.runTraining(imageClassifierTrainer, suite);
          return new ResponseEntity<String>("Training classifier in the background", HttpStatus.OK);
         }
     
-    //  @PostMapping(path="/detector", produces = "application/json")
-    // public ResponseEntity<String> trainDetectorAsync(@RequestBody @Valid TrainingSuite suite){ 
-    //      trainService.trainDetector(suite);
-    //      return new ResponseEntity<String>("Training detector in the background", HttpStatus.OK);
-    //     }
-
     }
