+++ service/src/main/java/net/softel/ai/classify/enums/NNet.java
@@ -3,5 +3,6 @@ package net.softel.ai.classify.enums;
 public enum NNet {
     UNDEFINED,
     RESNET_50,
-    MLP
+    YOLO_V5,
+    SSD
     }
