+++ service/src/main/java/net/softel/ai/classify/service/IPredict.java
@@ -8,8 +8,6 @@ import ai.djl.modality.Classifications;
 public interface IPredict {
     public String predictClass(PredictSuite suite);
     public String predictBest(PredictSuite suite);
-
-    // public Classifications predictClassification(PredictSuite predictSuite);
     }
