+++ service/src/main/java/net/softel/ai/classify/train/Training.java
@@ -1,3 +1,6 @@
+/**
+    All trainers must implement this intrefacce
+ */
 package net.softel.ai.classify.train;
 
 import net.softel.ai.classify.dto.TrainingSuite;
